refactor: replace Double with BigDecimal for monetary values

+++ src/main/java/com/project/sapaadu/dto/response/OrderItemResponse.java
@@ -3,11 +3,13 @@ package com.project.sapaadu.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class OrderItemResponse {
 
     private String itemName;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
 }
